API/requirements.txt
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
typing_extensions==4.6.3
